requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
canvasapi==3.3.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.4
django-allauth==65.3.0
django-filter==23.1
djangorestframework==3.15.2
idna==3.10
kombu==5.4.2
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
psycopg2-binary
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
pillow==12.0.0
django-celery-beat==2.7.0
django-anymail==12.0
django-debug-toolbar==5.1.0
brother_ql==0.9.4